integration_service/pyproject.toml
[project]
name = "znayka-integration-service"
version = "0.1.0"
description = "FastAPI wrapper around znayka-video script pipeline."
requires-python = ">=3.11"
dependencies = [
    "arq>=0.26.0",
    "fastapi>=0.115.0",
    "pydantic>=2.8.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
]
